# api/openapi.yaml
openapi: 3.1.0
info:
  title: Shubayr API
  version: 11.0.0
  description: >
    REST/JSON API contract for Shubayr — an online multi-section store
    (single-owner, white-label). This document is the SHARED CONTRACT between
    the backend (NestJS) and the clients (Next.js web, Flutter mobile/admin).
    It must stay in sync with infra/db/schema.sql. Changing an endpoint's shape
    is a reviewed PR that updates this file first.

    Access model v2 (breaking): every bearer token has `surface=admin|app`.
    Admin tokens come only from username/password login and staff operations
    require both `surface=admin` and the route's `x-permission`. App tokens come
    only from phone OTP, contain the server-resolved single app role, and never
    authorize admin operations. App clients must not send a role during OTP.
    Work roles (`delivery_agent`, `order_monitor`) cannot use purchase functions.

    Validation and errors: malformed or invalid request data returns HTTP 422
    with code `VALIDATION_FAILED` and field-level errors. All non-2xx JSON
    responses use the Error schema. PATCH distinguishes omission from null:
    omitted fields are unchanged, while null is accepted only for explicitly
    nullable fields and clears that field.

    Money policy: every money-bearing resource identifies its ISO currency.
    Amounts are stored exactly and displayed at that currency's configured
    precision (IQD 0, USD 2). The ledger preserves original-currency amount and
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /admin/currencies:
    get:
      tags: [Finance]
      summary: List configured currencies
      x-surface: admin
      x-permission: fx_rates.view
      responses:
        "200":
          description: Currencies
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Currency" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/currencies/{code}:
    patch:
      tags: [Finance]
      summary: Enable a currency or select the base before the first movement
      x-surface: admin
      x-permission: settings.manage
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, pattern: '^[A-Z]{3}$' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                enabled: { type: boolean }
                is_base: { type: boolean }
      responses:
        "200": { description: Updated currency, content: { application/json: { schema: { $ref: "#/components/schemas/Currency" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/exchange-rates:
    get:
      tags: [Finance]
      summary: List exchange-rate history newest first
      x-surface: admin
      x-permission: fx_rates.view
      parameters:
        - name: currency_code
          in: query
          schema: { type: string, pattern: '^[A-Z]{3}$' }
      responses:
        "200": { description: Rates, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ExchangeRate" } } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Finance]
      summary: Record an effective exchange rate
      description: Input is `1 foreign = X base`; basis 100 is normalized to basis 1 before storage. `effective_at` is an exact timestamp and a future-effective rate is never applied early.
      x-surface: admin
      x-permission: fx_rates.update
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ExchangeRateInput" } } }
      responses:
        "201": { description: Recorded rate, content: { application/json: { schema: { $ref: "#/components/schemas/ExchangeRate" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/exchange-rates/{code}/applicable:
    get:
      tags: [Finance]
      summary: Resolve the latest rate effective at an instant
      x-surface: admin
      x-permission: fx_rates.view
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, pattern: '^[A-Z]{3}$' }
        - name: at
          in: query
          required: true
          schema: { type: string, format: date-time }
      responses:
        "200": { description: Applicable rate with stale-today warning, content: { application/json: { schema: { $ref: "#/components/schemas/ExchangeRate" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ExchangeRateValidation" }
  /admin/exchange-rates/linked-price-preview:
    post:
      tags: [Finance]
      summary: Preview a pricing rate across every linked SKU
      description: Returns old/new prices and signed percentage changes plus a short-lived, actor-bound token. No rate or product price is changed.
      x-surface: admin
      x-permission: fx_rates.update
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LinkedPricePreviewInput" } } }
      responses:
        "201": { description: Linked-price preview, content: { application/json: { schema: { $ref: "#/components/schemas/LinkedPricePreview" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/exchange-rates/save-rate-only:
    post:
      tags: [Finance]
      summary: Save the previewed rate without publishing linked prices
      description: Linked SKUs keep their published price and are marked as awaiting this new rate.
      x-surface: admin
      x-permission: fx_rates.update
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LinkedPriceApplyInput" } } }
      responses:
        "201": { description: Rate saved, content: { application/json: { schema: { $ref: "#/components/schemas/LinkedPriceApplyResult" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/exchange-rates/publish-linked-prices:
    post:
      tags: [Finance]
      summary: Atomically save a rate and publish one linked-price version
      description: Refuses an expired or stale preview. Conversion and configured upward rounding occur before product discounts. If a resulting price is below protected cost, no rate or price is saved; a pending approval request is created for a different user to approve or reject.
      x-surface: admin
      x-permissions: [fx_rates.update, prices.publish_linked]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LinkedPriceApplyInput" } } }
      responses:
        "201": { description: Rate and linked prices published, content: { application/json: { schema: { $ref: "#/components/schemas/LinkedPriceApplyResult" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/price-publish-approvals/{id}/decision:
    post:
      tags: [Finance]
      summary: Approve or reject a pending below-cost linked-price publish
      description: The authenticated proposer is stored server-side and can never decide their own request.
      x-surface: admin
      x-permission: sell_below_cost.approve
      x-error-codes: [SEPARATION_OF_DUTIES_VIOLATION]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/PricePublishDecisionInput" } } }
      responses:
        "201": { description: Decided approval request, content: { application/json: { schema: { $ref: "#/components/schemas/PricePublishApproval" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/operations/{operationId}:
    get:
      tags: [Finance]
      summary: Get the caller's idempotent operation outcome
      x-surface: admin
      parameters:
        - name: operationId
          in: path
          required: true
          schema: { type: string, minLength: 8, maxLength: 128 }
      responses:
        "200": { description: Operation outcome, content: { application/json: { schema: { $ref: "#/components/schemas/OperationOutcome" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/drafts:
    get:
      tags: [Finance]
      summary: List the caller's financial document drafts
      x-surface: admin
      responses:
        "200": { description: Drafts, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/DocumentDraft" } } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/drafts/{documentType}:
    parameters:
      - name: documentType
        in: path
        required: true
        schema: { type: string, maxLength: 60 }
    get:
      tags: [Finance]
      summary: Get the caller's draft
      x-surface: admin
      responses:
        "200": { description: Draft, content: { application/json: { schema: { $ref: "#/components/schemas/DocumentDraft" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Finance]
      summary: Save the caller's draft without financial effect
      x-surface: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [payload]
              properties:
                payload: { type: object, additionalProperties: true }
      responses:
        "200": { description: Saved draft, content: { application/json: { schema: { $ref: "#/components/schemas/DocumentDraft" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
    delete:
      tags: [Finance]
      summary: Discard the caller's draft
      x-surface: admin
      responses:
        "204": { description: Draft discarded }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/ledger/entries:
    get:
      tags: [Finance]
      summary: List immutable journal entries
      x-surface: admin
      x-permission: ledger.view
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
        - { name: account_code, in: query, schema: { type: string } }
        - { name: source_type, in: query, schema: { type: string } }
        - { name: source_id, in: query, schema: { type: string, format: uuid } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200": { description: Journal page, content: { application/json: { schema: { $ref: "#/components/schemas/JournalEntryPage" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/ledger/entries/{id}:
    get:
      tags: [Finance]
      summary: Read one immutable journal entry with lines and reversal links
      x-surface: admin
      x-permission: ledger.view
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200": { description: Journal entry, content: { application/json: { schema: { $ref: "#/components/schemas/JournalEntry" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/ledger/entries/{id}/reversal:
    post:
      tags: [Finance]
      summary: Post the one allowed reversal for an immutable entry
      x-surface: admin
      x-permission: ledger.reverse
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AuditReason" } } } }
      responses:
        "201": { description: Reversal entry, content: { application/json: { schema: { $ref: "#/components/schemas/JournalEntry" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/PostingConflict" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/ledger/trial-balance:
    get:
      tags: [Finance]
      summary: Get trial balance
      x-surface: admin
      x-permission: ledger.view
      parameters:
        - { name: as_of, in: query, schema: { type: string, format: date } }
      responses:
        "200": { description: Trial balance, content: { application/json: { schema: { $ref: "#/components/schemas/TrialBalance" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/ledger/accounts/{code}/balance:
    get:
      tags: [Finance]
      summary: Get a ledger account balance
      x-surface: admin
      x-permission: ledger.view
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
        - { name: as_of, in: query, schema: { type: string, format: date } }
      responses:
        "200": { description: Account balance, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/cash-accounts:
    get:
      tags: [Finance]
      summary: List cash and bank accounts with ledger-derived balances
      x-surface: admin
      x-permission: cash_accounts.view
      responses:
        "200": { description: Accounts, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/CashAccount" } } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Finance]
      summary: Create a cash or bank account
      x-surface: admin
      x-permission: cash_accounts.manage
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CashAccountInput" } } } }
      responses:
        "201": { description: Created account, content: { application/json: { schema: { $ref: "#/components/schemas/CashAccount" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/cash-accounts/{id}:
    get:
      tags: [Finance]
      summary: Read a cash or bank account with its ledger-derived balance
      x-surface: admin
      x-permission: cash_accounts.view
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200": { description: Cash account, content: { application/json: { schema: { $ref: "#/components/schemas/CashAccount" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Finance]
      summary: Rename or activate a cash account
      x-surface: admin
      x-permission: cash_accounts.manage
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string, minLength: 2, maxLength: 160 }
                is_active: { type: boolean }
      responses:
        "200": { description: Updated account, content: { application/json: { schema: { $ref: "#/components/schemas/CashAccount" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Validation" }
    delete:
      tags: [Finance]
      summary: Delete an unused cash account; accounts with movements must be disabled
      x-surface: admin
      x-permission: cash_accounts.manage
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "204": { description: Unused account deleted }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/cash-accounts/{id}/opening-balance:
    post:
      tags: [Finance]
      summary: Post a cash account opening-balance document
      x-surface: admin
      x-permission: cash_accounts.manage
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/OpeningBalanceInput" } } } }
      responses:
        "201": { description: Posted opening balance, content: { application/json: { schema: { $ref: "#/components/schemas/FinancialDocument" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/PostingConflict" }
        "422": { $ref: "#/components/responses/ExchangeRateValidation" }
  /admin/cash-transfers:
    post:
      tags: [Finance]
      summary: Post a numbered same-currency transfer
      x-surface: admin
      x-permission: cash_accounts.manage
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CashTransferInput" } } } }
      responses:
        "201": { description: Posted transfer, content: { application/json: { schema: { $ref: "#/components/schemas/FinancialDocument" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/PostingConflict" }
        "422": { $ref: "#/components/responses/ExchangeRateValidation" }
  /admin/financial-documents/{id}:
    get:
      tags: [Finance]
      summary: Read one posted financial document by id
      description: Cash transfers and opening balances require either ledger.view or cash_accounts.view.
      x-surface: admin
      x-permissions-any: [ledger.view, cash_accounts.view]
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200": { description: Posted financial document, content: { application/json: { schema: { $ref: "#/components/schemas/FinancialDocument" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/accounting-periods:
    get:
      tags: [Finance]
      summary: List accounting periods and close history
      x-surface: admin
      x-permission: ledger.view
      responses:
        "200": { description: Periods, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/AccountingPeriod" } } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/accounting-periods/{month}/checklist:
    get:
      tags: [Finance]
      summary: Run the close checklist
      x-surface: admin
      x-permission: period.close
      parameters: [{ name: month, in: path, required: true, schema: { type: string, pattern: '^\d{4}-\d{2}$' } }]
      responses:
        "200": { description: Checklist, content: { application/json: { schema: { $ref: "#/components/schemas/PeriodChecklist" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/accounting-periods/{month}/close:
    post:
      tags: [Finance]
      summary: Close a period and save a sequenced trial-balance snapshot
      x-surface: admin
      x-permission: period.close
      parameters: [{ name: month, in: path, required: true, schema: { type: string, pattern: '^\d{4}-\d{2}$' } }]
      requestBody: { required: false, content: { application/json: { schema: { type: object, additionalProperties: false, properties: { reason: { type: string, maxLength: 500 } } } } } }
      responses:
        "201": { description: Period close, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/accounting-periods/{month}/reopen:
    post:
      tags: [Finance]
      summary: Reopen a closed period with a mandatory reason
      x-surface: admin
      x-permission: period.reopen
      parameters: [{ name: month, in: path, required: true, schema: { type: string, pattern: '^\d{4}-\d{2}$' } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AuditReason" } } } }
      responses:
        "201": { description: Reopened period, content: { application/json: { schema: { $ref: "#/components/schemas/AccountingPeriod" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }

  # ---------------- ADMIN: INVENTORY & COSTING ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
...
  responses:
...
    NotFound: { description: Not found (`NOT_FOUND`), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Conflict: { description: State conflict (`CONFLICT`), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    Validation: { description: "Invalid request data (`VALIDATION_FAILED`, HTTP 422)", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    ExchangeRateValidation: { description: "Invalid request data (`VALIDATION_FAILED`) or no applicable currency rate (`EXCHANGE_RATE_NOT_FOUND`), HTTP 422", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    Forbidden: { description: Missing permission (`FORBIDDEN`), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  schemas:
...
    Currency:
      type: object
      required: [code, name_ar, name_en, symbol, display_precision, is_base, enabled, created_at, updated_at]
      properties:
        code: { type: string, pattern: '^[A-Z]{3}$' }
        name_ar: { type: string }
        name_en: { type: string }
        symbol: { type: string }
        display_precision: { type: integer, minimum: 0, maximum: 6 }
        is_base: { type: boolean }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ExchangeRateInput:
      type: object
      additionalProperties: false
      required: [currency_code, rate, basis, effective_at, reason]
      properties:
        currency_code: { type: string, pattern: '^[A-Z]{3}$' }
        rate: { type: string, pattern: '^\d+(?:\.\d{1,10})?$' }
        basis: { type: integer, enum: [1, 100] }
        effective_at: { type: string, format: date-time, description: Exact instant from which the rate applies. }
        reason: { type: string, minLength: 3, maxLength: 500 }
    ExchangeRate:
      type: object
      properties:
        id: { type: string, format: uuid }
        currency_code: { type: string, pattern: '^[A-Z]{3}$' }
        base_currency_code: { type: string, pattern: '^[A-Z]{3}$' }
        rate: { $ref: "#/components/schemas/ExactDecimal" }
        effective_at: { type: [string, "null"], format: date-time }
        set_by: { type: string, format: uuid }
        reason: { type: string }
        created_at: { type: string, format: date-time }
        direction: { type: string }
        not_from_today: { type: boolean }
    LinkedPricePreviewInput:
      type: object
      additionalProperties: false
      required: [currency_code, rate, basis, effective_at, reason]
      properties:
        currency_code: { type: string, pattern: '^[A-Z]{3}$' }
        rate: { type: number, exclusiveMinimum: 0, multipleOf: 0.0000000001 }
        basis: { type: integer, enum: [1, 100] }
        effective_at: { type: string, format: date-time }
        reason: { type: string, maxLength: 500 }
...
    LinkedPricePreview:
      type: object
      additionalProperties: false
      required: [preview_token, expires_at, currency_code, old_rate, new_rate, linked_sku_count, rounding_multiple, items]
      properties:
        preview_token: { type: string, format: uuid }
        expires_at: { type: string, format: date-time }
        currency_code: { type: string, pattern: '^[A-Z]{3}$' }
        old_rate: { type: [number, "null"] }
        new_rate: { type: number }
        linked_sku_count: { type: integer, minimum: 0 }
        rounding_multiple: { type: number, minimum: 0 }
        items: { type: array, items: { $ref: "#/components/schemas/LinkedPricePreviewItem" } }
    LinkedPriceApplyInput:
      type: object
      additionalProperties: false
      required: [preview_token]
      properties:
        preview_token: { type: string, format: uuid }
        below_cost_override_reason: { type: [string, "null"], maxLength: 500 }
    LinkedPriceApplyResult:
      type: object
      additionalProperties: false
      required: [mode, exchange_rate_id, price_version_id, approval_request_id, linked_sku_count]
      properties:
        mode: { type: string, enum: [rate_only, published, pending_approval] }
        exchange_rate_id: { type: [string, "null"], format: uuid }
        price_version_id: { type: [string, "null"], format: uuid }
        approval_request_id: { type: [string, "null"], format: uuid }
        linked_sku_count: { type: integer, minimum: 0 }
    PricePublishDecisionInput:
      type: object
      additionalProperties: false
      required: [decision, reason]
      properties:
        decision: { type: string, enum: [approve, reject] }
        reason: { type: string, minLength: 3, maxLength: 500 }
...